refactor: `OpenApi` internal module
